detect code and data sections automatically
The sections that are considered code or data are based on
the section flags instead on a hardcoded list of sections.
Similarly, .bss sections are detected based on the flags, not on
the name.
We still have a hardcoded list of special section in ddisasm to be
used by heuristics but it is a "whitelist" rather than a "blacklist".