Allow using '-' as an output file name to print to stdout.
The --ir, --json, and --asm options all now support printing to stdout
when the file name specified is - or when no file name is specified.

This required us to switch printing status messages from stdout to
stderr instead. It also requires us to be careful when printing
binary output to stdout because of line feed translation when in text
mode.

This also fixes an issue with JSON printing where it would also
automatically disassemble the output.