Consider stack variables and break stratification between value_reg and value_reg_edge
This commit contains several important and related changes.
- First it starts tracking flow-insensitive stack variables for the value analysis.
- Second, it breaks stratification between value_reg and value_reg_edge.
This means that we can start using value_reg for finding jump tables (and in other places).
To do that, we need to splicitly determine locations where registers take a value
that is uknown or that we cannot model, instead of relying on negation value_reg_edge to come up
with these locations.

These two changes allow us to detect additional jump tables in which the address of the
jump table is temporarily stored in the stack.