Fix fallthrough cfg_edge
(1) nop
    nop
(2) instr1
    instr2
(3) instr3

The block A in nop_block(A) and next_refined_block_in_section(A,_) are
not always the same block: the A in nop_block is the original refined
block (e.g., (1) or (2)), whereas the A in next_refined_block_in_section can be a merged one (e.g., (1)+(2)), especially when (2) doesn't have any label (no predecessor).
In such case, with the previous predicate
cfg_edge(Src,Dest,"false","false","fallthrough"):-
     nop_block(Src),
     cfg_edge(_,Src,_,_,_),
     next_refined_block_in_section(Src,Dest).
the edge from (2) to (3) can be incorrectly recognized as fallthrough.
To avoid such cases, this change checks if there's any non-nop block
between (1) and (3).