Revert the previous change for fallthrough
In the previous commit, may_fallthrough takes the last instruction of a
block instead of block's control instruction. However, this change may
be not correct. The unit tests passed, but the change not necessarilly
affects them, but affects the gtirb IR, which was not thoroughly tested.
For now, revert the change.