Use improved noreturn for def-used CFG
During use-def and value analysis, the idea of the CFG is held by the
last_block relation. This relies almost entirely on the pre-code
inference view of the CFG, which has very primitive detection of PLT
entries.

This adds a no-return relation that uses more complete PLT blocks, and
utilizes this relation to refine the def-used last_block relation.