Fix regression for adrp-add through stack
The restriction of detecting stack loads to only cases where the loaded
address for an address broke some cases of adrp-add split-load detection
across stack loads, since some cases did not use the loaded value
directly as an address (e.g., when the register is passed as an argument
into a dynamically-linked library function).

This adds an additional rule for stack_var_used_for_address to detect
where a value is used as an address and then saved to the stack. In
combination with a change that ensures `adrp` instructions are always
considered used-for-address, this ensures that these cases are detected.

This does come at a performance cost. This additional
stack_var_used_for_address rule adds a runtime of 33.3 sec to the
bins/aarch64.ubuntu20.bitcoin.clang.O0.nopie.strip.elf binary, making it
the second-worst rule. However, this is still better than foregoing the
used-for-address restriction on stack_load_store_pair altogether, which
adds so many additional stack_load_store_pair and value_reg_edge that
they would more than offset the cost of this restriction.