Populate constant values directly into value_reg
It does not make sense for a constant value to feed into
value_reg_edge, since it's more of a "node" than a "edge" when
considering the value analysis as a graph. Further, the constant
value_reg_edge are not used directly anywhere, other than to create
value_reg with a 1:1 correspondance, so it is a bit redudant.

In libpcre2-8.so with AMD64 gcc -O2, this makes some small improvements
in runtime and relation size:

* value_reg_edge: 178k, 1.63s -> 176k, 1.44s
* value_reg: 3.53M, 65.0s -> 3.53M, 63.9s

This also allows changing the type used in value_reg_edge from
"reg_nullable" to "register", which improves type safety.