Disable var-tracking-assignments for datalog code
The var-tracking-assignments option greatly increases the required
memory to build the datalog code; often on our CI runners, the option
fails (presumably due to OOM) and gcc retries without the option. This
retry substantially increases the total build time.

The var-tracking-assignments option improves debug information, but
debug information for the datalog code is unlikely to be used.