Apply "fits" points to both code and data blocks
The "fits" heuristic implies that a block is likely real, but doesn't
actually imply anything about the type of the block. Thus, we should
apply this to both code and data blocks, since a fitting data block is
just as valid as a fitting code block (i.e., if a code block and data
block both fit equally well, the "fit" rules should not tie-break in
favor of code).

These heuristics were written before separate data blocks existed, so
applying them to only code blocks was likely an oversight.