Copy files into docker image from local repository
Previously, the ddisasm Docker image built ddisasm by cloning a fresh
copy of the repository during the image build. However, this could
result in unintended caching when the same branch is built on a newer
commit (as the CI does for main).

Building from the checked-out files ensures the intended versions of the
ddisasm source is used to build the image.