Resolve ARM jump table conflicts with block_points
Previously, a single data block candidate was always generated for the
ARM jump table, with a few heuristics to decide if a direct_jump into
the jump table entries was real or not.

Now, we split the data block candidate at potential direct jumps and use
block_points to resolve conflicts between the data and code blocks.

In addition, we assign points to jump table entry blocks that align with
a jump table target.